allow for resource attributes that aren't mapped
We would like to rely on the Datadog mapping of resource attributes, but
are unable to supply custom resource attributes as tags.

At the moment there are only two options:

* Use `ResourceAttributesAsTags` which will map the resource attributes
  to Datadog specific tags, but ignore any resource attributes that do
  not get mapped.
* Do not use `ResourceAttributesAsTags`, and allow for custom tags
  through resource attributes, but lose the ability to map tags
  correctly. This means that we need to reimplement the default mappings
  provided by the exporter elsewhere in the otel pipeline.

This change updates the Datadog Exporter so that we can leverage the
default mappings, while also providing custom resource attributes.

A common use case is that we rely on Kubernetes labels to provide
service ownership information in the form of a `team` label. We are
relying on the k8s_tagger to extract this label from the Kubernetes
resource and pass it along.